backend/requirements.txt
livekit==1.0.12
openai==1.101.0
python-dotenv==1.0.0
